% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_selection.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Helper functions for data aggregation and summary statistics}
\usage{
aggregate_data(data_filtered, metric)
}
\description{
Helper functions for data aggregation and summary statistics
}
